fix: pass order items to order
