+++ src/adapter/driver/routes/doc/order.ts
@@ -649,94 +649,3 @@ export const SwaggerCreateOrder = {
 		},
 	},
 };
-
-export const SwaggerAddItemToCart = {
-	schema: {
-		summary: 'Add order item to cart',
-		description: 'Add order item to cart and return created order item',
-		tags: ['Order'],
-		params: { id: { type: 'string', format: 'uuid' } },
-		body: {
-			productId: { type: 'string' },
-			quantity: { type: 'number' },
-			details: { type: 'string' },
-		},
-		response: {
-			201: {
-				description: 'Item added to cart successfully',
-				type: 'object',
-				properties: {
-					id: {
-						type: 'string',
-					},
-					orderId: {
-						type: 'string',
-					},
-					productId: {
-						type: 'string',
-					},
-					quantity: {
-						type: 'number',
-					},
-					value: {
-						type: 'number',
-					},
-					details: {
-						type: 'string',
-						format: 'nullable',
-					},
-					createdAt: {
-						type: 'string',
-						format: 'date-time',
-					},
-					updatedAt: {
-						type: 'string',
-						format: 'date-time',
-					},
-				},
-			},
-			400: {
-				description: 'Bad request',
-				type: 'object',
-				properties: {
-					path: {
-						type: 'string',
-					},
-					status: {
-						type: 'number',
-					},
-					message: {
-						type: 'string',
-					},
-					details: {
-						type: 'array',
-						items: {
-							type: 'string',
-						},
-					},
-				},
-			},
-			500: {
-				description: 'Server error',
-				type: 'object',
-				properties: {
-					path: {
-						type: 'string',
-					},
-					status: {
-						type: 'string',
-					},
-					message: {
-						type: 'string',
-					},
-					details: {
-						type: 'array',
-						items: {
-							type: 'string',
-						},
-					},
-				},
-			},
-		},
-	},
-};
